+++ app/src/feature/admin/allMembers/index/table/PendingActivationMemberRow.tsx
@@ -1,7 +1,7 @@
 import type { PendingActivationMember } from "@/core/domains/member/pendingActivationMember";
 import type { FC } from "react";
 import { ListItem } from "../list/ListItem";
-import { Popup } from "../popup/Popup2";
+import { Popup } from "../popup/Popup";
 import { ThreeDotsIcon } from "../threeDotsIcon/ThreeDotsIcon";
 import type { OnClickBan } from "./type";
